Add documentation for all structures in the code base
