+++ Cargo.toml
@@ -21,7 +21,7 @@ tracing = "0.1.40"
 
 [workspace.lints.rust]
 missing_debug_implementations = "warn"
-# missing_docs = "warn"
+missing_docs = "warn"
 non_ascii_idents = "deny"
 trivial_casts = "warn"
 trivial_numeric_casts = "warn"
@@ -41,7 +41,7 @@ allow_attributes_without_reason = "deny"
 # cargo_common_metadata = "warn"
 clone_on_ref_ptr = "warn"
 default_union_representation = "deny"
-# missing_docs_in_private_items = "warn"
+missing_docs_in_private_items = "warn"
 mixed_read_write_in_expression = "deny"
 # multiple_crate_versions = "warn"
 negative_feature_names = "deny"
