+++ crates/huub/src/model.rs
@@ -1,3 +1,6 @@
+//! Module that defines the main data structures to formulate, i.e. model, a
+//! problem instance to be solved.
+
 pub(crate) mod bool;
 pub(crate) mod branching;
 pub(crate) mod constraint;
@@ -34,20 +37,30 @@ use crate::{
 };
 
 #[derive(Debug, Default)]
+/// A formulation of a problem instance in terms of decisions and constraints.
 pub struct Model {
+	/// A base [`Cnf`] object that contains pure Boolean parts of the problem.
 	pub(crate) cnf: Cnf,
+	/// An list of branching strategies that will be used by created [`Solver`]
+	/// instances to be used in order to make search decisions.
 	branchings: Vec<Branching>,
+	/// A list of constraints that have been added to the model.
 	constraints: Vec<Constraint>,
+	/// The definitions of the integer variables that have been created.
 	int_vars: Vec<IntVarDef>,
+	/// A queue of indexes of constraints that need to be propagated.
 	prop_queue: VecDeque<usize>,
+	/// A flag for each constraint whether it has been enqueued for propagation.
 	enqueued: Vec<bool>,
 }
 
 impl Model {
+	/// Create a new Boolean variable.
 	pub fn new_bool_var(&mut self) -> BoolView {
 		BoolView::Lit(self.cnf.new_lit())
 	}
 
+	/// Create `len` new Boolean variables.
 	pub fn new_bool_vars(&mut self, len: usize) -> Vec<BoolView> {
 		self.cnf
 			.new_var_range(len)
@@ -55,12 +68,14 @@ impl Model {
 			.collect()
 	}
 
+	/// Create a new integer variable with the given domain.
 	pub fn new_int_var(&mut self, domain: RangeList<i64>) -> IntView {
 		let iv = IntVar(self.int_vars.len() as u32);
 		self.int_vars.push(IntVarDef::with_domain(domain));
 		IntView::Var(iv)
 	}
 
+	/// Create `len` new integer variables with the given domain.
 	pub fn new_int_vars(&mut self, len: usize, domain: RangeList<i64>) -> Vec<IntVar> {
 		let iv = IntVar(self.int_vars.len() as u32);
 		self.int_vars
@@ -68,6 +83,13 @@ impl Model {
 		(0..len).map(|i| IntVar(iv.0 + i as u32)).collect()
 	}
 
+	/// Process the model to create a [`Solver`] instance that can be used to
+	/// solve it.
+	///
+	/// This method will return a [`Solver`] instance in addition to a
+	/// [`VariableMap`], which can be used to map from [`ModelView`] to
+	/// [`SolverView`]. If an error occurs during the reformulation process, or if
+	/// it is found to be trivially unsatisfiable, then an error will be returned.
 	pub fn to_solver<Oracle: PropagatingSolver<Engine>>(
 		&mut self,
 		config: &InitConfig,
@@ -150,6 +172,7 @@ impl Model {
 		Ok((slv, map))
 	}
 
+	/// Enqueue constraint that has index `constraint` to the propagation queue.
 	fn enqueue(&mut self, constraint: usize) {
 		if !self.enqueued[constraint] {
 			self.prop_queue.push_back(constraint);
@@ -194,15 +217,20 @@ impl ClauseDatabase for Model {
 }
 
 #[derive(Clone, Debug, PartialEq, Eq, Hash)]
+/// Reference to a decision in a [`Model`].
 pub enum ModelView {
+	/// Reference to a Boolean decision.
 	Bool(BoolView),
+	/// Reference to an integer decision.
 	Int(IntView),
 }
+
 impl From<IntView> for ModelView {
 	fn from(value: IntView) -> Self {
 		Self::Int(value)
 	}
 }
+
 impl From<BoolView> for ModelView {
 	fn from(value: BoolView) -> Self {
 		Self::Bool(value)
