+++ crates/huub/src/propagator.rs
@@ -1,3 +1,5 @@
+//! Module containing the definitions for propagators and their implementations.
+
 pub(crate) mod all_different_int;
 pub(crate) mod array_int_minimum;
 pub(crate) mod array_var_int_element;
@@ -16,7 +18,7 @@ use std::fmt::Debug;
 use pindakaas::Lit as RawLit;
 
 use crate::{
-	actions::{explanation::ExplanationActions, propagation::PropagationActions},
+	actions::{ExplanationActions, PropagationActions},
 	propagator::conflict::Conflict,
 	solver::{
 		engine::{solving_context::SolvingContext, State},
@@ -25,6 +27,16 @@ use crate::{
 	Conjunction,
 };
 
+/// A trait for a propagator that is called during the search process to filter
+/// the domains of decision variables, and detect inconsistencies.
+///
+/// Implementations of the propagator trait must be able to explain changes to
+/// domains of decision variables as a conjunction of literals that imply the
+/// change. If these explanations are too expensive to compute during
+/// propagation, then the propagator can delay giving the explanation using
+/// [`PropagationActions::deferred_reason`]. If the explanation is needed, then
+/// the propagation engine will revert the state of the solver and call
+/// [`Propagator::explain`] to receive the explanation.
 pub(crate) trait Propagator<P: PropagationActions, E: ExplanationActions>:
 	Debug + DynPropClone
 {
@@ -43,8 +55,12 @@ pub(crate) trait Propagator<P: PropagationActions, E: ExplanationActions>:
 	/// propagated.
 	///
 	/// The method is called with the data that was passed to the
-	/// `deferred_reason` method, and the literal that was propagated. If the
-	/// `lit` argument is `None`, then the reason was used to explain `false`.
+	/// [`PropagationActions::deferred_reason`] method, and the literal that was
+	/// propagated. If the `lit` argument is `None`, then the reason was used to
+	/// explain `false`.
+	///
+	/// The state of the solver is reverted to the state before the propagation of
+	/// the `lit` to be explained.
 	fn explain(&mut self, actions: &mut E, lit: Option<RawLit>, data: u64) -> Conjunction {
 		let _ = actions;
 		let _ = lit;
@@ -54,14 +70,20 @@ pub(crate) trait Propagator<P: PropagationActions, E: ExplanationActions>:
 	}
 }
 
+/// A trait to allow the cloning of boxed propagators.
+///
+/// This trait allows us to implement [`Clone`] for [`BoxedPropagator`].
 pub(crate) trait DynPropClone {
+	/// Clone the object and store it as a boxed trait object.
 	fn clone_dyn_prop(&self) -> BoxedPropagator;
 }
+
 impl<P: for<'a> Propagator<SolvingContext<'a>, State> + Clone + 'static> DynPropClone for P {
 	fn clone_dyn_prop(&self) -> BoxedPropagator {
 		Box::new(self.clone())
 	}
 }
+
 impl Clone for BoxedPropagator {
 	fn clone(&self) -> BoxedPropagator {
 		self.clone_dyn_prop()
