+++ crates/huub/src/solver/engine/bool_to_int.rs
@@ -1,3 +1,6 @@
+//! Module containing structures for tracking the relationships between Boolean
+//! variables and integer variables.
+
 use std::collections::HashMap;
 
 use pindakaas::{Var as RawVar, VarRange};
@@ -5,12 +8,19 @@ use pindakaas::{Var as RawVar, VarRange};
 use crate::{solver::engine::int_var::IntVarRef, LitMeaning};
 
 #[derive(Default, Debug, Clone, PartialEq, Eq)]
+/// A mapping of Boolean variables to integer variables of which they represent
+/// conditions.
 pub(crate) struct BoolToIntMap {
+	/// The mapping of eagerly created Boolean variables to the integer variables.
 	eager: Vec<(VarRange, IntVarRef)>,
+	/// The mapping of lazily created Boolean variables to the integer variables
+	/// and their meanings.
 	lazy: HashMap<RawVar, (IntVarRef, LitMeaning)>,
 }
 
 impl BoolToIntMap {
+	/// Insert a range of Boolean variables to map them to the integer variable
+	/// for which they are eagerly created to represent conditions for.
 	pub(crate) fn insert_eager(&mut self, range: VarRange, var: IntVarRef) {
 		if range.is_empty() {
 			return;
@@ -22,11 +32,16 @@ impl BoolToIntMap {
 		panic!("Literal Mapping not added in the correct order")
 	}
 
+	/// Insert a mapping of a lazily created Boolean variable to the integer
+	/// variable and the meaning of the literal on the integer variable.
 	pub(crate) fn insert_lazy(&mut self, var: RawVar, iv: IntVarRef, lit: LitMeaning) {
 		let x = self.lazy.insert(var, (iv, lit));
 		debug_assert_eq!(x, None, "lazy literal already exists");
 	}
 
+	/// Return the integer variable the given Boolean variable represents a
+	/// condition for, if any. If the Boolean variable was lazily created, then
+	/// also return the [`LitMeaning`] of the literal.
 	pub(crate) fn get(&self, var: RawVar) -> Option<(IntVarRef, Option<LitMeaning>)> {
 		let is_eager = self
 			.eager
